+++ ximpol/srcmodel/config/test_uniform_disk.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 #
-# Copyright (C) 2015, the ximpol team.
+# Copyright (C) 2015--2016, the ximpol team.
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU GengReral Public License as published by
@@ -22,14 +22,16 @@ import numpy
 from ximpol.srcmodel.roi import xUniformDisk, xROIModel
 from ximpol.srcmodel.spectrum import power_law, constant
 
+
 ROI_MODEL = xROIModel(10., 10.)
 
-source1 = xUniformDisk(name='source1', ra=10., dec=10., radius=0.01)
-source1.spectrum = power_law(10., 2.)
-source1.polarization_degree = constant(0.5)
-source1.polarization_angle = constant(numpy.radians(65.))
+energy_spectrum = power_law(10., 2.)
+polarization_degree = constant(0.5)
+polarization_angle = constant(numpy.radians(65.))
+disk = xUniformDisk('disk', 10., 10., 0.01, energy_spectrum,
+                    polarization_degree, polarization_angle)
 
-ROI_MODEL.add_source(source1)
+ROI_MODEL.add_source(disk)
 
 
 if __name__ == '__main__':